requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain
langchain-core
langchain-community
langchain-groq
langchain-ollama
langchain-chroma
chromadb
sentence-transformers
PyMuPDF
psycopg2-binary
sqlalchemy
